refactor: simplify label count endpoint
