+++ api/app/api/api_v1/endpoints/reddit/label.py
@@ -1,10 +1,10 @@
 from collections import Counter
 
-from fastapi import APIRouter, HTTPException
-from starlette.status import HTTP_200_OK, HTTP_404_NOT_FOUND
+from fastapi import APIRouter
+from starlette.status import HTTP_200_OK
 
 from app.models.post import Post
-from app.schemas.post import PostCount
+from app.schemas.reddit import PostCount
 
 router = APIRouter()
 
@@ -12,12 +12,4 @@ router = APIRouter()
 @router.get("/", response_model=PostCount, status_code=HTTP_200_OK)
 async def get_label_counts():
     """Check how many posts are in DB for each label."""
-    posts = await Post.objects.all()
-    if not posts:
-        raise HTTPException(HTTP_404_NOT_FOUND, "No posts found in DB")
-
-    counter = Counter()
-    for post in posts:
-        counter[post.label] += 1
-
-    return dict(counter)
+    return Counter((post.label for post in await Post.objects.all()))
